chore: configure ssh to work with Teleport servers (#4)
This change is part of the overall project configuration. It's aim is to
have a shareable ssh configuration that can be versioned controlled
together with the rest of the repository contents. This will also make
it easier to onboard new users of the playbooks.

To achieve this, the aim is to make most of the inventory if not all accessible
through Teleport and as such, included here is also a way to configure the
user's computer to allow Ansible to work with Teleport.