chore: add generic teleport config during installation (#21)
This adds a [Teleport](https://goteleport.com/docs) config during installation and provides mechanism to override specific parts for the [configuration](https://goteleport.com/docs/reference/config) to suite the installation context. In short, with this role and proper overriding of the configuration, Teleport can be installed and configured to provide any or all of it's services.

It should be noted though, that by default, the installed Teleport will not be configured to do anything useful and in fact, the Teleport `systemd` service is not even started. It's up to the clients/users of this role to provide the proper configuration overrides as per the installation context.