feat: update KenyaEMR to the latest version (#26)
This is an update of the KenyaEMR installation role to use the latest version of KenyaEMR which is version 18.4.2 as of this commit.
Changes have been made to the bundled database so that it is compatible
with the latest KenyaEMR version. The default OpenMRS Platform WAR file deployed by the `openmrs` role has also been changed from the version 2.3.3 downloaded from the official OpenMRS site to the openMRS WAR file build by Palledium to make it easier to deploy KenyaEMR. However the role has also being updated to still support the former behaviour when that is required.

Included are other minor fixes and improvements.