feat: add Eclipse Jetty installation role (#35)
Add a role to install [Eclipse Jetty Server](https://www.eclipse.org/jetty) on an Ubuntu 20.04 LTS(and above) host.
This installation is more suitable for the root installation of Jetty as
defined by the `JETTY_HOME` property. For runtime specific
installations(defined by the `JETTY_BASE` property), a seperate role
will be added in the future to address that.

Included also is the addition of a new invertory group, `test_servers` and
also other minor fixes and improvements to existing roles.